Commit 1
Cloned the repository down and studied all of the files.   Added a command to load the init. Attached CSS and Javascript files to the HTML document.  Started writing commands for two functions.  Added a for loop inside the second command. Declared multiple global variables at the top of the Javascript fie